Add stencil geometry classes.
This includes ABCs for sky-coordinate and pixel-coordinate stencils,
and implementations for circle and great-circle polygons.

Test coverage for polygons is just okay, but doing better is a lot
harder, and I'd prefer to wait to see if we scrap this for delegation
to AST before working any harder on it.