Add SkyStencil property to get a unique hash/encoding for filenames.
There isn't a lot of thought behind any of the decisions that went into
this implementation (e.g. hash algorithm, whether to use a hash
algorithm at all).  If IVOA has any kind of standard or precedent for
a compact hash of the stencils we are supposed to support, that would
be better, as would anything that did have some thought behind it.