Add FITS write methods and high-level interface.
This FITS write implementation uses ResourcePath and afw writeFits
methods directly, instead of delegating to Butler or even Datastore.
This will probably need to be revisited in the future, but aside from
the sidestepping philosophical and design problems with delegating to
Butler for these, this works around the more immediate problem that we
can't write custom metadata for anything other than Exposure objects
via the Butler, because we don't have a way to forward to the optional
'metadata' parameter in the afw writeFits methods for Image, Mask, and
MaskedImage.